Added context_corpus#each_document
Added a function to iterate over each synthetic 'document' in a context_corpus.
For a sparsevector file of the form:

   foo 0 1 1 1 2 2
   bar 0 1 3 1

the function produces four documents, one per dimension, with the words
appearing in each document determined by the words which appear in the context
of the dimension's original word. For the previous example, we get the following
documents:

   0: foo bar
   1: foo
   2: foo foo
   3: bar